+++ test/api/user/create-v1.js
@@ -9,7 +9,7 @@ module.exports.fields = {
 	password: String
 }
 
-module.exports.output = {}
+module.exports.outFields = {}
 
 module.exports.handler = function (body, success) {
 	assert(process.domain.runInfo.requestId.length === 24)